requirements.txt
astroid==2.2.5
Django==2.2.5
django-cleanup==4.0.0
django-crispy-forms==1.7.2
django-jet==1.0.8
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==6.1.0
psycopg2==2.8.3
pylint==2.3.1
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
